transform byte to int
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C52527

+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Data.SQL/SupplierRateDataManager.cs
@@ -142,7 +142,7 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
                 CurrencyId = GetReaderValue<int?>(reader, "CurrencyID"),
                 Rate = GetReaderValue<decimal>(reader, "Rate"),
                 RateTypeId = GetReaderValue<int?>(reader, "RateTypeID"),
-                RateChange = (RateChangeType)GetReaderValue<byte>(reader, "Change"),
+                RateChange = (RateChangeType)GetReaderValue<int>(reader, "Change"),
                 BED = GetReaderValue<DateTime>(reader, "BED"),
                 EED = GetReaderValue<DateTime?>(reader, "EED")
             };
